backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sse-starlette==1.8.2
aiohttp==3.9.5
dnspython==2.6.1
python-nmap==0.7.1
python-whois==0.9.5
